requirements.txt
Django==2.0.1
djangorestframework==3.7.7
django-cors-headers==2.1.0
django-filter==1.1.0
django-jsoneditor==0.0.12
future==0.16.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
urllib3==1.22
gunicorn==19.7.1
